Update tensor shape assertion in float_tensor.rs
Changed the value tested in the assertion for the shape of the tensor, in the 'float_tensor' file. Previously, it was checking if the shape is a vector with a single element of '1'. Now, it verifies whether the shape is an empty vector as it should be for scalar values.